# packages/web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Royal Battalion API
  description: |
    REST API for the Royal Battalion gaming community management platform.
    Handles user authentication (Discord OAuth), whitelist management,
    ticket/prospect tracking, match history, server administration,
    and Discord bot integration.

    ## Versioning
    All REST endpoints are versioned and served under the `/v1` prefix, which is
    baked into the `servers[].url` base below (so the paths documented here are
    relative to `/v1`, e.g. `/users` resolves to `/v1/users`).

    A few root-level, NON-versioned endpoints are exempt (external / infra
    contracts that must not move) and declare their own root `servers` override:
    - `GET /health` -- infra/Railway health probe.
    - `GET /admins.cfg` -- fetched by the Squad game server at a fixed URL.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via the `/auth/sync` endpoint.
    The flow is:
    1. User authenticates with Discord OAuth on the frontend
    2. Frontend sends the Discord access token to `POST /auth/sync`
    3. API returns a JWT (valid 4 hours) containing userId and permissions
    4. Include the JWT in all subsequent requests as `Authorization: Bearer <token>`

    ## Permissions
    Access is controlled by a granular RBAC system with 19 permissions.
    Users with the `admin` permission bypass all permission checks.

    ## Rate Limiting
    - Global: 200 requests/minute per IP
...

...
  version: 1.0.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /matches/public:
    get:
      tags: [Matches]
      summary: Get recent public matches (paginated)
      description: Returns non-hidden matches, most recent first. No authentication required.
      security: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated public matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PaginatedMatches"
  /matches:
    get:
      tags: [Matches]
      summary: List all matches (paginated)
      description: Returns all matches including hidden ones, most recent first.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PaginatedMatches"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:matches
  /matches/{id}:
    patch:
      tags: [Matches]
      summary: Update match
      description: Partially update match details (date, map, layer, result, etc.).
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date-time
                map:
                  type: string
                layer:
                  type: string
                result:
                  type: string
                server:
                  type: string
                vodUrl:
                  type: string
                  nullable: true
                hidden:
                  type: boolean
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_Match"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Match not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:matches
    delete:
      tags: [Matches]
      summary: Delete match
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Match not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:matches
  /matches/resync:
    post:
      tags: [Matches]
      summary: Resync matches from SquadJS
      description: Triggers a full resync of match data from the SquadJS database.
      responses:
        "200":
          description: Resync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      resynced:
                        type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:matches

  # ── Servers ───────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/sync
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page number
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 50
      description: Page size (server-capped)
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NoContent:
      description: Success (no content)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    Paginated:
      type: object
      description: Standard pagination envelope for growing collections.
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNext:
          type: boolean
...
    ApiResponse_PaginatedMatches:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          allOf:
            - $ref: "#/components/schemas/Paginated"
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/Match"
...
    MatchTeam:
      type: object
      properties:
        faction:
          type: string
        factionFull:
          type: string
        result:
          type: string
          enum: [WIN, LOSS, DRAW]
    MatchPlayer:
      type: object
      properties:
        name:
          type: string
        steamId:
          type: string
        squad:
          type: string
        role:
          type: string
        kills:
          type: integer
        deaths:
          type: integer
        revives:
          type: integer
        teamkills:
          type: integer
        isSquadLeader:
          type: boolean
    MatchDetail:
      type: object
      properties:
        duration:
          type: string
        players:
          type: integer
        team1:
          $ref: "#/components/schemas/MatchTeam"
        team2:
          $ref: "#/components/schemas/MatchTeam"
        team1Players:
          type: array
          items:
            $ref: "#/components/schemas/MatchPlayer"
        team2Players:
          type: array
          items:
            $ref: "#/components/schemas/MatchPlayer"
    Match:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          format: date-time
        map:
          type: string
        layer:
          type: string
        result:
          type: string
        server:
          type: string
        vodUrl:
          type: string
          nullable: true
        hidden:
          type: boolean
        matchDetail:
          oneOf:
            - $ref: "#/components/schemas/MatchDetail"
            - type: "null"
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiResponse_Match:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Match"
...